backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-telegram-bot==21.4
channels==4.1.0

